+++ Source/Rendering/CrRenderingForwardDeclarations.h
@@ -6,7 +6,7 @@
 
 namespace cr3d
 {
-	namespace DataFormat { enum T : uint8_t; }
+	namespace DataFormat { enum T : uint32_t; }
 	enum class TextureType : uint8_t;
 	using TextureUsageFlags = uint32_t;
 	enum class SampleCount :uint8_t;
